+++ project/Http4sPlugin.scala
@@ -300,7 +300,7 @@ object Http4sPlugin extends AutoPlugin {
     // error-prone merge conflicts in the dependencies below.
     val argonaut = "6.2.5"
     val asyncHttpClient = "2.10.5"
-    val blaze = "0.14.14"
+    val blaze = "0.14.15"
     val boopickle = "1.3.3"
     val cats = "2.3.1"
     val catsEffect = "2.3.1"